+++ packages/gamut/src/ConnectedForm/ConnectedInputs/ConnectedRadioGroupInput.tsx
@@ -6,13 +6,19 @@ import { ConnectedRadioGroupInputProps } from './types';
 export const ConnectedRadioGroupInput: React.FC<ConnectedRadioGroupInputProps> = ({
   name,
   options,
+  disabled,
   ...rest
 }) => {
   return (
     <ConnectedRadioGroup name={name} {...rest}>
       {options.map((elem) => {
         return (
-          <ConnectedRadio name={name} key={`${name}-${elem.value}`} {...elem} />
+          <ConnectedRadio
+            name={name}
+            key={`${name}-${elem.value}`}
+            disabled={disabled}
+            {...elem}
+          />
         );
       })}
     </ConnectedRadioGroup>
