--- packages/styleguide/stories/Organisms/Forms/ConnectedForm.examples.tsx
@@ -1,493 +1,0 @@
-import {
-  Box,
-  Card,
-  ConnectedCheckbox,
-  ConnectedInput,
-  ConnectedRadioGroupInput,
-  ConnectedSelect,
-  ConnectedTextArea,
-  FlexBox,
-  FloatingCard,
-  SubmitButton,
-  Text,
-  useConnectedForm,
-} from '@codecademy/gamut';
-import {
-  MiniArrowRightIcon,
-  TerminalIcon,
-  ViewIcon,
-} from '@codecademy/gamut-icons';
-import { Keyhole } from '@codecademy/gamut-illustrations';
-import { DotDense } from '@codecademy/gamut-patterns';
-import { Background } from '@codecademy/gamut-styles';
-import { action } from '@storybook/addon-actions';
-import React, { useState } from 'react';
-
-const INeedSomeSpace: React.FC = ({ children }) => {
-  return <Box m={16}>{children}</Box>;
-};
-
-export const RadioGroupForm = () => {
-  const {
-    ConnectedFormGroup,
-    ConnectedForm,
-    connectedFormProps,
-  } = useConnectedForm({
-    defaultValues: {
-      radioGroup: undefined,
-    },
-    validationRules: {
-      radioGroup: {
-        required: 'please fill this out.',
-      },
-    },
-  });
-
-  return (
-    <ConnectedForm
-      onSubmit={(values) => {
-        action('Form Submitted')(values);
-      }}
-      resetOnSubmit
-      {...connectedFormProps}
-    >
-      <Text>I have something important to tell you...</Text>
-      <SubmitButton variant="secondary" m={8}>
-        submit this form.
-      </SubmitButton>
-      <ConnectedFormGroup
-        name="radioGroup"
-        label="cool radiogroup bruh"
-        field={{
-          component: ConnectedRadioGroupInput,
-          options: [
-            { label: 'one', value: 'one' },
-            { label: 'two', value: 'two' },
-            { label: 'zero', value: 'zero' },
-          ],
-        }}
-      />
-    </ConnectedForm>
-  );
-};
-
-export const RadioWatchExample = () => {
-  const [showRadio, setShowRadio] = useState(false);
-
-  const handleWatch = (currentFormValues: { checkbox: boolean }) => {
-    return currentFormValues.checkbox
-      ? setShowRadio(true)
-      : setShowRadio(false);
-  };
-
-  const {
-    ConnectedFormGroup,
-    ConnectedForm,
-    connectedFormProps,
-  } = useConnectedForm({
-    defaultValues: {
-      checkbox: false,
-      radioGroup: undefined,
-    },
-    validationRules: {
-      radioGroup: {
-        required: 'please fill this out.',
-      },
-    },
-    watchedFields: {
-      fields: ['checkbox'],
-      watchHandler: handleWatch,
-    },
-  });
-
-  return (
-    <ConnectedForm onSubmit={() => {}} resetOnSubmit {...connectedFormProps}>
-      <Text>I have something important to tell you...</Text>
-      <SubmitButton variant="secondary" m={8}>
-        submit this form.
-      </SubmitButton>
-      <ConnectedFormGroup
-        name="checkbox"
-        label="cool radiogroup bruh"
-        field={{
-          component: ConnectedCheckbox,
-          label: 'check me to view the radio',
-        }}
-      />
-      {showRadio && (
-        <ConnectedFormGroup
-          name="radioGroup"
-          label="cool radiogroup bruh"
-          field={{
-            component: ConnectedRadioGroupInput,
-            options: [
-              { label: 'one', value: 'one' },
-              { label: 'two', value: 'two' },
-              { label: 'zero', value: 'zero' },
-            ],
-          }}
-        />
-      )}
-    </ConnectedForm>
-  );
-};
-
-export const ConnectedFormExample = () => {
-  const {
-    ConnectedFormGroup,
-    ConnectedForm,
-    connectedFormProps,
-  } = useConnectedForm({
-    defaultValues: {
-      checkboxField: false,
-      selectField: 'zero',
-      inputField: '',
-      radioGroupField: undefined,
-      textAreaField: '',
-    },
-    validationRules: {
-      checkboxField: { required: 'you need to check this.' },
-      selectField: {
-        pattern: {
-          value: /^(?:(?!zero).)*$/,
-          message: 'literally anything but zero',
-        },
-      },
-      inputField: { required: 'we need this info, bud' },
-      radioGroupField: { required: 'we need this info too, bud' },
-      textAreaField: {
-        required: 'you just have to fill out the whole thing, okay?',
-      },
-    },
-  });
-  return (
-    <ConnectedForm
-      {...connectedFormProps}
-      onSubmit={(values) => console.log(values)}
-      resetOnSubmit
-      height="60rem"
-      display="flex"
-      flexDirection="column"
-      alignItems="center"
-      justifyContent="space-between"
-    >
-      <Text>I have something important to tell you...</Text>
-      <SubmitButton variant="secondary" m={8}>
-        submit this form.
-      </SubmitButton>
-      <ConnectedFormGroup
-        name="checkboxField"
-        label="checkbox field"
-        field={{
-          component: ConnectedCheckbox,
-          label: 'check it ouuut',
-        }}
-      />
-      <ConnectedFormGroup
-        name="selectField"
-        label="select field"
-        field={{
-          component: ConnectedSelect,
-          options: ['one', 'two', 'zero'],
-        }}
-      />
-      <ConnectedFormGroup
-        name="inputField"
-        label="input field"
-        field={{
-          component: ConnectedInput,
-          icon: TerminalIcon,
-        }}
-      />
-      <ConnectedFormGroup
-        name="radioGroupField"
-        label="radio group field"
-        field={{
-          component: ConnectedRadioGroupInput,
-          options: [
-            {
-              label: (
-                <>
-                  <MiniArrowRightIcon mr={4} /> a
-                </>
-              ),
-              value: 'a',
-            },
-            {
-              label: (
-                <>
-                  <MiniArrowRightIcon mr={4} /> b
-                </>
-              ),
-              value: 'b',
-            },
-            {
-              label: (
-                <>
-                  <MiniArrowRightIcon mr={4} /> c
-                </>
-              ),
-              value: 'c',
-            },
-            {
-              label: (
-                <>
-                  <MiniArrowRightIcon mr={4} />
-                  zilch
-                </>
-              ),
-              value: 'zilch',
-            },
-          ],
-        }}
-      />
-      <ConnectedFormGroup
-        name="textAreaField"
-        label="text area field"
-        field={{
-          component: ConnectedTextArea,
-        }}
-      />
-    </ConnectedForm>
-  );
-};
-
-export const FancyConnectedFormExample: React.FC = () => {
-  const {
-    ConnectedFormGroup,
-    ConnectedForm,
-    connectedFormProps,
-  } = useConnectedForm({
-    defaultValues: {
-      exampleRadioArea0: undefined,
-      select0: undefined,
-      select1: 'two',
-      select2: undefined,
-      input1: 'howdy.',
-      exampleRadioArea1: '5',
-      exampleIcon: 'hey',
-      check0: true,
-      check1: false,
-      check2: true,
-      exampleTextArea: '',
-    },
-    validationRules: {
-      exampleRadioArea0: {
-        required: 'you must make a decision',
-      },
-      select0: { required: 'hey..' },
-      select1: {
-        pattern: {
-          value: /^(?:(?!zero).)*$/,
-          message: 'zero to hero',
-        },
-      },
-      select2: {
-        pattern: {
-          value: /^(?:(?!one).)*$/,
-          message: 'anything but one',
-        },
-      },
-      input1: {
-        pattern: {
-          value: /^(?:(?!zero).)*$/,
-          message: 'zero to hero',
-        },
-      },
-      check0: { required: 'check it...' },
-      check1: { required: 'check it...' },
-      check2: { required: 'check it...' },
-    },
-  });
-
-  return (
-    <Background bg="paleGreen" border={1} p={32} m={32} borderRadius="10px">
-      <ConnectedForm
-        onSubmit={(values) => values}
-        resetOnSubmit
-        display="grid"
-        gridTemplateRows="auto"
-        gridTemplateAreas="'illo illo header header'
-    'illo illo rightBar rightBar'
-    'main main main main'
-    'footer footer footer footer'"
-        {...connectedFormProps}
-      >
-        <Box gridArea="header" display="flex" justifyContent="flex-end">
-          <Box>
-            <FloatingCard beak="bottom-right" pattern={DotDense}>
-              <SubmitButton variant="secondary" m={32}>
-                dont forget to submit
-              </SubmitButton>
-            </FloatingCard>
-          </Box>
-        </Box>
-        <Box
-          gridArea="rightBar"
-          bg="palePink"
-          border={1}
-          p={16}
-          m={24}
-          borderRadius="10px"
-        >
-          <INeedSomeSpace>
-            <ConnectedFormGroup
-              name="exampleRadioArea0"
-              label="pls answer"
-              field={{
-                component: ConnectedRadioGroupInput,
-                options: [
-                  { label: 'one', value: 'first' },
-                  { label: 'two', value: 'two' },
-                  { label: 'three', value: 'three' },
-                  { label: 'zilch', value: 'zero' },
-                ],
-              }}
-            />
-          </INeedSomeSpace>
-          <INeedSomeSpace>
-            <ConnectedFormGroup
-              name="select0"
-              label="pls... answer..."
-              field={{
-                component: ConnectedSelect,
-                options: ['one', 'two'],
-              }}
-            />
-          </INeedSomeSpace>
-        </Box>
-        <Box
-          gridArea="main"
-          bg="palePink"
-          border={1}
-          p={24}
-          my={12}
-          mx={24}
-          borderRadius="10px"
-        >
-          <INeedSomeSpace>
-            <ConnectedFormGroup
-              name="select1"
-              label="render select ok"
-              field={{
-                component: ConnectedSelect,
-                options: ['zero', 'two', 'three'],
-              }}
-            />
-            <ConnectedFormGroup
-              name="select2"
-              label="render select ok tré"
-              field={{
-                component: ConnectedSelect,
-                options: ['zero', 'one', 'two'],
-              }}
-            />
-            <ConnectedFormGroup
-              name="input1"
-              label="render input ok"
-              field={{
-                component: ConnectedInput,
-              }}
-            />
-          </INeedSomeSpace>
-          <FlexBox m={12}>
-            <ConnectedFormGroup
-              name="exampleRadioArea1"
-              label="render radio ok"
-              field={{
-                component: ConnectedRadioGroupInput,
-                options: [
-                  { label: 'one', value: '1' },
-                  { label: 'two', value: '2' },
-                  { label: 'three', value: '2' },
-                  { label: 'four', value: '4' },
-                  { label: 'five', value: '5' },
-                ],
-              }}
-            />
-            <ConnectedFormGroup
-              name="exampleIcon"
-              label="write text in here"
-              spacing="tight"
-              errorType="initial"
-              field={{
-                component: ConnectedInput,
-                icon: ViewIcon,
-              }}
-            />
-          </FlexBox>
-        </Box>
-        <Card shadow="medium" gridArea="illo">
-          <Keyhole />
-          <Background
-            bg="navy-900"
-            borderRadius="10px"
-            display="flex"
-            justifyContent="center"
-            p={16}
-          >
-            <Text as="code" justifySelf="center" truncate="fade">
-              we are always watching......
-            </Text>
-          </Background>
-        </Card>
-
-        <Background
-          gridArea="footer"
-          bg="navy"
-          p={16}
-          mt={12}
-          borderRadius="10px"
-        >
-          <Card shadow="medium">
-            <ConnectedFormGroup
-              name="check0"
-              label="hi"
-              spacing="tight"
-              errorType="initial"
-              field={{
-                component: ConnectedCheckbox,
-                label: 'check me out',
-                spacing: 'tight',
-              }}
-            />
-            <ConnectedFormGroup
-              name="check1"
-              label="hi"
-              spacing="tight"
-              errorType="initial"
-              field={{
-                component: ConnectedCheckbox,
-                label: 'check me out',
-                spacing: 'tight',
-              }}
-            />
-            <ConnectedFormGroup
-              name="check2"
-              label="hi"
-              spacing="tight"
-              errorType="initial"
-              field={{
-                component: ConnectedCheckbox,
-                label: 'check me out',
-                spacing: 'tight',
-              }}
-            />
-          </Card>
-          <INeedSomeSpace>
-            <ConnectedFormGroup
-              name="exampleTextArea"
-              label="write text in here"
-              spacing="tight"
-              errorType="initial"
-              field={{
-                component: ConnectedTextArea,
-                placeholder: 'pls write your doctoral thesis ty',
-              }}
-            />
-          </INeedSomeSpace>
-        </Background>
-      </ConnectedForm>
-    </Background>
-  );
-};
