+++ packages/styleguide/stories/Organisms/Forms/ConnectedForm/ConnectedForm.examples.tsx
@@ -1,0 +1,304 @@
+import {
+  ConnectedCheckbox,
+  ConnectedFormGroupProps,
+  ConnectedFormProps,
+  ConnectedInput,
+  ConnectedRadioGroupInput,
+  ConnectedSelect,
+  ConnectedTextArea,
+  SubmitButton,
+  Text,
+  useConnectedForm,
+} from '@codecademy/gamut';
+import { MiniArrowRightIcon, TerminalIcon } from '@codecademy/gamut-icons';
+import { action } from '@storybook/addon-actions';
+import React, { useState } from 'react';
+
+export const FormGroupStates = () => {
+  const {
+    ConnectedFormGroup,
+    ConnectedForm,
+    connectedFormProps,
+  } = useConnectedForm({
+    defaultValues: {
+      radioGroup: undefined,
+      radioGroupError: undefined,
+      radioGroupDisabled: undefined,
+    },
+    validationRules: {
+      radioGroup: {
+        required: 'please fill this out.',
+      },
+    },
+  });
+
+  return (
+    <ConnectedForm
+      display="flex"
+      mt={12}
+      onSubmit={(values) => {
+        action('Form Submitted')(values);
+      }}
+      resetOnSubmit
+      {...connectedFormProps}
+    >
+      <ConnectedFormGroup
+        name="radioGroup"
+        label="default state"
+        size="large"
+        field={{
+          component: ConnectedRadioGroupInput,
+          options: [
+            { label: 'one', value: 'one' },
+            { label: 'two', value: 'two' },
+            { label: 'zero', value: 'zero' },
+          ],
+        }}
+      />
+      <ConnectedFormGroup
+        name="radioGroupError"
+        label="error state"
+        customError="custom error"
+        size="large"
+        field={{
+          component: ConnectedRadioGroupInput,
+          options: [
+            { label: 'one', value: 'one' },
+            { label: 'two', value: 'two' },
+            { label: 'zero', value: 'zero' },
+          ],
+        }}
+      />
+      <ConnectedFormGroup
+        name="radioGroupDisabled"
+        label="disabled state"
+        size="large"
+        disabled
+        field={{
+          component: ConnectedRadioGroupInput,
+          options: [
+            { label: 'one', value: 'one' },
+            { label: 'two', value: 'two' },
+            { label: 'zero', value: 'zero' },
+          ],
+        }}
+      />
+    </ConnectedForm>
+  );
+};
+
+export const RadioWatchExample = () => {
+  const [showRadio, setShowRadio] = useState(false);
+  const [lastFormValues, setLastFormValues] = useState({
+    checkbox: 'not submitted',
+    radioGroup: 'not submitted',
+  });
+
+  const handleWatch = ({ checkbox }: { checkbox?: undefined | boolean }) => {
+    return checkbox ? setShowRadio(true) : setShowRadio(false);
+  };
+
+  const {
+    ConnectedFormGroup,
+    ConnectedForm,
+    connectedFormProps,
+  } = useConnectedForm({
+    defaultValues: {
+      checkbox: false,
+      radioGroup: undefined,
+    },
+    validationRules: {
+      radioGroup: {
+        required: 'please fill this out.',
+      },
+    },
+    watchedFields: {
+      fields: ['checkbox', 'radioGroup'],
+      watchHandler: handleWatch,
+    },
+  });
+
+  return (
+    <ConnectedForm
+      onSubmit={({ checkbox, radioGroup }) => {
+        setLastFormValues({
+          checkbox: `${checkbox}`,
+          radioGroup: `${radioGroup}`,
+        });
+      }}
+      resetOnSubmit
+      {...connectedFormProps}
+    >
+      <Text as="code" lineHeight="base" mb={24}>
+        your last answers were {lastFormValues.checkbox} and{' '}
+        {lastFormValues.radioGroup}!
+      </Text>
+      <ConnectedFormGroup
+        name="checkbox"
+        label="a special checkbox"
+        spacing="tight"
+        field={{
+          component: ConnectedCheckbox,
+          label: 'check me to view the radio',
+          spacing: 'tight',
+        }}
+      />
+      {showRadio && (
+        <ConnectedFormGroup
+          name="radioGroup"
+          label="cool radiogroup bruh"
+          field={{
+            component: ConnectedRadioGroupInput,
+            options: [
+              { label: 'one', value: 'one' },
+              { label: 'two', value: 'two' },
+              { label: 'zero', value: 'zero' },
+            ],
+          }}
+        />
+      )}
+      <SubmitButton m={8}>submit the form, please</SubmitButton>
+    </ConnectedForm>
+  );
+};
+
+type ConnectedFormPlayground = {
+  connectedForm: Omit<
+    ConnectedFormProps<any>,
+    'defaultValue' | 'validationRules' | 'watchedFields' | 'onSubmit'
+  >;
+  connectedFormGroup: Omit<
+    ConnectedFormGroupProps<any>,
+    'name' | 'label' | 'field'
+  >;
+};
+
+export const ConnectedFormPlayground: React.FC<ConnectedFormPlayground> = ({
+  connectedForm,
+  connectedFormGroup,
+}) => {
+  const {
+    ConnectedFormGroup,
+    ConnectedForm,
+    connectedFormProps,
+  } = useConnectedForm({
+    defaultValues: {
+      checkboxField: false,
+      selectField: 'zero',
+      inputField: '',
+      radioGroupField: undefined,
+      textAreaField: '',
+    },
+    validationRules: {
+      checkboxField: { required: 'you need to check this.' },
+      selectField: {
+        pattern: {
+          value: /^(?:(?!zero).)*$/,
+          message: 'literally anything but zero',
+        },
+      },
+      inputField: { required: 'we need this info, bud' },
+      radioGroupField: { required: 'we need this info too, bud' },
+      textAreaField: {
+        required: 'you just have to fill out the whole thing, okay?',
+      },
+    },
+  });
+  return (
+    <ConnectedForm
+      onSubmit={(values) => {
+        action('Form Submitted')(values);
+      }}
+      display="flex"
+      flexDirection="column"
+      alignItems="center"
+      justifyContent="space-between"
+      {...connectedFormProps}
+      {...connectedForm}
+    >
+      <Text>
+        I have something important to tell you...
+        <SubmitButton m={8}>submit this</SubmitButton>
+      </Text>
+
+      <ConnectedFormGroup
+        name="checkboxField"
+        label="checkbox field"
+        field={{
+          component: ConnectedCheckbox,
+          label: 'check it ouuut',
+        }}
+        {...connectedFormGroup}
+      />
+      <ConnectedFormGroup
+        name="selectField"
+        label="select field"
+        field={{
+          component: ConnectedSelect,
+          options: ['one', 'two', 'zero'],
+        }}
+        {...connectedFormGroup}
+      />
+      <ConnectedFormGroup
+        name="inputField"
+        label="input field"
+        field={{
+          component: ConnectedInput,
+          icon: TerminalIcon,
+        }}
+        {...connectedFormGroup}
+      />
+      <ConnectedFormGroup
+        name="radioGroupField"
+        label="radio group field"
+        field={{
+          component: ConnectedRadioGroupInput,
+          options: [
+            {
+              label: (
+                <>
+                  <MiniArrowRightIcon mr={4} /> a
+                </>
+              ),
+              value: 'a',
+            },
+            {
+              label: (
+                <>
+                  <MiniArrowRightIcon mr={4} /> b
+                </>
+              ),
+              value: 'b',
+            },
+            {
+              label: (
+                <>
+                  <MiniArrowRightIcon mr={4} /> c
+                </>
+              ),
+              value: 'c',
+            },
+            {
+              label: (
+                <>
+                  <MiniArrowRightIcon mr={4} />
+                  zilch
+                </>
+              ),
+              value: 'zilch',
+            },
+          ],
+        }}
+        {...connectedFormGroup}
+      />
+      <ConnectedFormGroup
+        name="textAreaField"
+        label="text area field"
+        field={{
+          component: ConnectedTextArea,
+        }}
+        {...connectedFormGroup}
+      />
+    </ConnectedForm>
+  );
+};
